[機能開発] versionサブコマンドの作成 change: config-createrをコンソールモードでビルドするように変更 change: Wallpaper Registoryをコンソールモードでビルドするように変更 add: versionサブコマンド作成 improve: Wallpaper Registoryのutf16PtrFromStringをgolang.org/x/sys/windows.StringToUTF16Ptrに変更
