+++ build.bat
@@ -1,7 +1,9 @@
-go get github.com/google/go-licenses
-go-licenses save ./... --force --save_path build/WallpaperChanger-Windows/licenses
-go build -o build/WallpaperChanger-Windows/ConfigCreater.exe -ldflags "-s -w -H=windowsgui" -trimpath ./cmd/config-creater
-go build -o build/WallpaperChanger-Windows/WallpaperChanger.exe -ldflags "-s -w" -trimpath ./cmd/wallpaper-changer
-go build -o build/WallpaperChanger-Windows/WallpaperChanger-no-console.exe -ldflags "-s -w -H=windowsgui" -trimpath ./cmd/wallpaper-changer
-go build -o build/WallpaperChanger-Windows/WallpaperRegistry.exe -ldflags "-s -w -H=windowsgui" -trimpath ./cmd/wallpaper-registry
-COPY README-windows.txt build\WallpaperChanger-Windows\README.txt
+CHCP 65001
+
+SET /P version=バージョンを入力: 
+go install github.com/google/go-licenses@latest
+go-licenses save ./... --force --save_path build/WallpaperChanger/licenses
+go build -o build/WallpaperChanger/ConfigCreater.exe -ldflags "-s -w -X=github.com/ryo-kagawa/WallpaperChanger/subcommand.name=ConfigCreater.exe -X=github.com/ryo-kagawa/WallpaperChanger/subcommand.version=%version%" -trimpath ./cmd/config-creater
+go build -o build/WallpaperChanger/WallpaperChanger.exe -ldflags "-s -w -H=windowsgui -X=github.com/ryo-kagawa/WallpaperChanger/subcommand.name=WallpaperChanger.exe -X=github.com/ryo-kagawa/WallpaperChanger/subcommand.version=%version%" -trimpath ./cmd/wallpaper-changer
+go build -o build/WallpaperChanger/WallpaperRegistry.exe -ldflags "-s -w -X=github.com/ryo-kagawa/WallpaperChanger/subcommand.name=WallpaperRegistry.exe -X=github.com/ryo-kagawa/WallpaperChanger/subcommand.version=%version%" -trimpath ./cmd/wallpaper-registry
+COPY README-windows.txt build\WallpaperChanger\README.txt
