+++ cmd/wallpaper-registry/main.go
@@ -3,46 +3,21 @@ package main
 import (
 	"fmt"
 	"os"
-	"strings"
-	"syscall"
 
-	"golang.org/x/sys/windows"
-	"golang.org/x/sys/windows/registry"
+	"github.com/ryo-kagawa/WallpaperChanger/subcommand"
+	"github.com/ryo-kagawa/go-utils/commandline"
 )
 
 func main() {
-	if !windows.GetCurrentProcessToken().IsElevated() {
-		utf16PtrFromString := func(value string) *uint16 {
-			ret, _ := syscall.UTF16PtrFromString(value)
-			return ret
-		}
-		verb := "runas"
-		exe, _ := os.Executable()
-		args := strings.Join(os.Args[1:], " ")
-		cwd, _ := os.Getwd()
-
-		const SW_SHOWNORMAL int32 = 1
-
-		err := windows.ShellExecute(
-			0,
-			utf16PtrFromString(verb),
-			utf16PtrFromString(exe),
-			utf16PtrFromString(args),
-			utf16PtrFromString(cwd),
-			SW_SHOWNORMAL,
-		)
-		if err != nil {
-			fmt.Println(err)
-		}
-		return
+	result, err := commandline.Execute(
+		Command{},
+		os.Args[1:],
+		subcommand.Version{},
+	)
+	if result != "" {
+		fmt.Fprint(os.Stdout, result)
 	}
-	key, _, err := registry.CreateKey(registry.CURRENT_USER, `Control Panel\Desktop`, registry.SET_VALUE|registry.CREATE_SUB_KEY)
 	if err != nil {
-		panic(err)
-	}
-	defer key.Close()
-	err = key.SetDWordValue("JPEGImportQuality", 100)
-	if err != nil {
-		panic(err)
+		fmt.Fprint(os.Stderr, err)
 	}
 }
