+++ go.mod
@@ -3,6 +3,7 @@ module github.com/ryo-kagawa/WallpaperChanger
 go 1.23.0
 
 require (
+	github.com/ryo-kagawa/go-utils v1.1.0
 	golang.org/x/image v0.19.0
 	golang.org/x/sys v0.24.0
 	gopkg.in/yaml.v3 v3.0.1
